+++ AdyenComponents/Issuer List/IssuerListComponent.swift
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2020 Adyen N.V.
+// Copyright (c) 2021 Adyen N.V.
 //
 // This file is open source and available under the MIT license. See the LICENSE file for more info.
 //
@@ -10,7 +10,7 @@ import UIKit
 
 /// A generic component for "issuer-based" payment methods, such as iDEAL and MOLPay.
 /// This component will provide a list in which the user can select their issuer.
-public final class IssuerListComponent: PaymentComponent, PresentableComponent {
+public final class IssuerListComponent: PaymentComponent, PresentableComponent, LoadingComponent {
     
     /// The issuer list payment method.
     public let paymentMethod: PaymentMethod
@@ -37,7 +37,7 @@ public final class IssuerListComponent: PaymentComponent, PresentableComponent {
     // MARK: - Presentable Component Protocol
     
     public var viewController: UIViewController {
-        return listViewController
+        listViewController
     }
     
     public func stopLoading(withSuccess success: Bool, completion: (() -> Void)?) {
